+++ src/main/java/survivalblock/crossbow_scoping/mixin/crossbow/TypedActionResultAccessor.java
@@ -1,4 +1,4 @@
-//? if <1.21.2 {
+//? if <=1.21.1 {
 /*package survivalblock.crossbow_scoping.mixin.crossbow;
 
 import net.minecraft.world.InteractionResultHolder;
@@ -6,6 +6,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
+@SuppressWarnings("UnusedMixin")
 @Mixin(InteractionResultHolder.class)
 public interface TypedActionResultAccessor<T> {