+++ src/main/java/survivalblock/crossbow_scoping/mixin/crossbow/client/InGameHudMixin.java
@@ -1,21 +1,31 @@
 package survivalblock.crossbow_scoping.mixin.crossbow.client;
 
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
-import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.gui.Gui;
+import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
+import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.Shadow;
 
 @Mixin(Gui.class)
-public class InGameHudMixin {
+public abstract class InGameHudMixin {
 
-    @WrapOperation(method = "renderItemHotbar", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/player/Player;getOffhandItem()Lnet/minecraft/world/item/ItemStack;"))
-    private ItemStack useRealCrossbow(Player instance, Operation<ItemStack> original) {
-        instance.crossbow_scoping$setAttacking(true);
-        ItemStack crossbow = original.call(instance);
-        instance.crossbow_scoping$setAttacking(false);
-        return crossbow;
+    @Shadow
+    @Nullable
+    protected abstract Player getCameraPlayer();
+
+    @WrapMethod(method = "renderItemHotbar")
+    private void useRealCrossbows(GuiGraphics guiGraphics, DeltaTracker deltaTracker, Operation<Void> original) {
+        Player player = this.getCameraPlayer();
+        if (player == null) {
+            original.call(guiGraphics, deltaTracker);
+            return;
+        }
+        player.crossbow_scoping$setAttacking(true);
+        original.call(guiGraphics, deltaTracker);
+        player.crossbow_scoping$setAttacking(false);
     }
 }