+++ src/main/java/survivalblock/crossbow_scoping/mixin/crossbow/client/ItemModelManagerMixin.java
@@ -1,15 +1,14 @@
+//? if >1.21.1 {
 package survivalblock.crossbow_scoping.mixin.crossbow.client;
 
 import net.minecraft.client.renderer.item.ItemModelResolver;
 import net.minecraft.client.renderer.item.ItemStackRenderState;
-import net.minecraft.util.Unit;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemDisplayContext;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -47,3 +46,4 @@ public class ItemModelManagerMixin {
         }
     }
 }
+//?}